Refactoring related to indexing features
